move some old tests
